refactor: Move fromKeyForIndexScan to same file
We have fromKeyForIndexScan and fromKeyForIndexScanInternal. Move them
next to each other.
